pyproject.toml
[project]
name = "eagleowl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["google-adk>=0.5.0"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 100
